requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.0
pillow==12.1.0
PyQt6==6.10.1
PyQt6-Charts==6.10.0
PyQt6-Charts-Qt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pyserial==3.5
requests==2.32.5
urllib3==2.6.2
websocket-client==1.9.0
